+++ src/spec/ParticleSpec.ts
@@ -25,6 +25,7 @@ function flushWebGLCanvasTo2D(source: HTMLCanvasElement): HTMLCanvasElement {
 
 describe('A particle', () => {
   let engine: ex.Engine;
+  let scene: ex.Scene;
   let texture: ex.Texture;
   beforeEach(() => {
     jasmine.addMatchers(ExcaliburMatchers);
@@ -37,6 +38,9 @@ describe('A particle', () => {
       },
       ['use-webgl']
     );
+    scene = new ex.Scene();
+    engine.addScene('root', scene);
+    engine.start();
 
     texture = new ex.Texture('base/src/spec/images/SpriteFontSpec/SpriteFont.png', true);
   });